EventBus is down. Long live CatnipShardImpl.
