+++ src/main/java/com/mewna/catnip/shard/manager/ShardManager.java
@@ -28,10 +28,8 @@
 package com.mewna.catnip.shard.manager;
 
 import com.mewna.catnip.Catnip;
-import com.mewna.catnip.shard.LifecycleState;
-import io.reactivex.Single;
+import com.mewna.catnip.shard.CatnipShard;
 
-import javax.annotation.CheckReturnValue;
 import javax.annotation.Nonnegative;
 import javax.annotation.Nonnull;
 import java.util.Collection;
@@ -63,7 +61,7 @@ public interface ShardManager {
     
     /**
      * @return The shard conditions for this shard manager instance.
-     *
+     *request
      * @see ShardCondition
      */
     List<ShardCondition> conditions();
@@ -85,68 +83,14 @@ public interface ShardManager {
     void addToConnectQueue(@Nonnegative int shard);
     
     /**
-     * Fetches the trace from the given shard.
+     * Fetches a shard.
      *
-     * @param shard The shard to fetch the trace from.
+     * @param shard The id of the shard to fetch.
      *
-     * @return The shard's trace.
+     * @return The shard, if it exists, otherwise it throws a {@link NullPointerException}.
      */
     @Nonnull
-    Single<List<String>> trace(@Nonnegative int shard);
-    
-    /**
-     * Return the shard's computed gateway latency, ie. the time it takes for
-     * the shard to send a heartbeat to Discord and get a response.
-     *
-     * @param shard The shard ID to get gateway latency for.
-     *
-     * @return The shard's computed gateway latency.
-     */
-    @Nonnull
-    Single<Long> latency(@Nonnegative int shard);
-    
-    /**
-     * Checks whether or not the shard with the given ID is currently connected
-     * to the websocket gateway. This is done as a boolean because - at least
-     * for now - there's only 2 meaningful states: connected, and queued to be
-     * connected.
-     * <p>
-     * If the shard id does <strong>not</strong> exist (ex. not controlled by
-     * this shard manager instance), then the future will FAIL, and this case
-     * must also be handled.
-     * <p>
-     * This is fetched <em>asynchronously</em>.
-     *
-     * @param id The id of the shard to get connection status for.
-     *
-     * @return Whether or not the shard with the given id is currently
-     * connected to the websocket gateway.
-     */
-    @Nonnull
-    @CheckReturnValue
-    Single<Boolean> isConnected(@Nonnegative int id);
-    
-    /**
-     * Get the lifecycle state for the current shard. This provides more
-     * meaningful info than {@link #isConnected(int)}, because it provides more
-     * granular info about the shard's state. However, it does not
-     * differentiate between "connected" or not per se; it is up to the
-     * end-user to determine whether or not the lifecycle state is actually a
-     * state of being "connected."
-     * <p>
-     * If the shard id does <strong>not</strong> exist (ex. not controlled by
-     * this shard manager instance), then the future will FAIL, and this case
-     * must also be handled.
-     * <p>
-     * This is fetched <em>asynchronously</em>.
-     *
-     * @param id The id of the shard to get the lifecycle state for.
-     *
-     * @return The lifecycle state for the shard with the given id.
-     */
-    @Nonnull
-    @CheckReturnValue
-    Single<LifecycleState> shardState(@Nonnegative int id);
+    CatnipShard shard(@Nonnegative int shard);
     
     /**
      * @return The catnip instance this shard manager is for.
